Add projects.yaml support
The projects.yaml lists all of the "official" StarlingX projects
(the ones actually owned by StarlingX and not upstream forks).
It is in YAML format to be machine-readable and useful for a
number of tools.

The first such use is to generate project-specific pages in the
governance documentation listing the projects, tech leads and
project leads. The current content is basically taken from
the existing wiki pages and is lacking in certain areas.
Please review it carefully.

projects.py is based on the OpenStack governance teams.py.

Introduce hacking and switch to flake8 for Python linting,
and fix the pep8 errors in members.py.

Change-Id: I514de8d574df8e60bb488974f2796e2c41010363
Signed-off-by: Dean Troyer <dtroyer@gmail.com>